Update Podcast Script
Fireside RSS stopped putting individual episode art in each item in the
feed, so the script was breaking. This uses the channel (postcast) album
art as the episode art for now, the page looks poor with all the same
cover art over and over, but it at least works. I also added the section for
season 9 and the quick links to jump down the page for more recent
seasons that had been missed.

We can revisit to see if something can be changed in Fireside, or was
changed in fireside, to cause this issue. If the image comes back, the
script should pick it up still.